condense utilities into single module
