+++ README.md
@@ -81,7 +81,7 @@ Run the unit tests with:
 ```bash
 pytest -v
 ```
-The test suite exercises the application factory and a wide range of utilities in the `utils` package, including diagram generation helpers.
+The test suite exercises the application factory and a wide range of utilities in the `utils` module, including diagram generation helpers.
 
 ### Custom configuration
 Set the ``CONFIG_PATH`` environment variable to load an alternative
