+++ utils.py
@@ -1,0 +1,540 @@
+"""Unified utility module for Rules Central."""
+
+from __future__ import annotations
+
+import os
+import re
+import json
+import uuid
+import logging
+from pathlib import Path
+from dataclasses import dataclass, asdict
+import dataclasses
+from datetime import datetime, timezone, timedelta
+from typing import Any, Dict, Iterable, List
+
+from werkzeug.utils import secure_filename
+from flask import current_app
+from flask_login import current_user
+
+from config import Config
+
+LOGGER = logging.getLogger(__name__)
+
+# ---------------------------------------------------------------------------
+# JSON helpers
+# ---------------------------------------------------------------------------
+
+def remove_all_quotes(obj: Any) -> Any:
+    """Recursively remove all double quotes from strings in ``obj``."""
+    if isinstance(obj, str):
+        return obj.replace('"', "")
+    if isinstance(obj, dict):
+        return {k: remove_all_quotes(v) for k, v in obj.items()}
+    if isinstance(obj, list):
+        return [remove_all_quotes(item) for item in obj]
+    return obj
+
+
+def add_missing_guids_if_needed(rules: Iterable[Dict[str, Any]]) -> None:
+    """Recursively add GUIDs to rules if they are missing."""
+    def recurse(rule):
+        if "RuleGUID" not in rule:
+            rule["RuleGUID"] = str(uuid.uuid4())
+        for child in rule.get("Children", []):
+            recurse(child)
+        for action in rule.get("Actions", []):
+            for child in action.get("ChildRules", []):
+                recurse(child)
+
+    for r in rules:
+        recurse(r)
+
+
+def load_and_sanitize_json(filepath: str | Path) -> list[Dict[str, Any]] | None:
+    """Load JSON from ``filepath`` and sanitize unexpected fields."""
+    try:
+        with open(filepath, "r", encoding="utf-8") as file:
+            content = file.read()
+            if content.lstrip().startswith("<"):
+                raise ValueError("Uploaded file appears to be HTML, not JSON.")
+            file.seek(0)
+            data = json.load(file)
+
+        if isinstance(data, dict) and "rules" in data:
+            data = data["rules"]
+        if not isinstance(data, list):
+            raise ValueError("JSON data must be a list of rules.")
+
+        allowed_fields = {
+            "RuleGUID",
+            "RuleName",
+            "Children",
+            "Actions",
+            "Attributes",
+            "ParentGUID",
+            "ParentActionIndex",
+            "Container",
+            "FunctionName",
+            "RootName",
+        }
+
+        def sanitize_rule(rule: Dict[str, Any]) -> Dict[str, Any]:
+            cleaned = {k: v for k, v in rule.items() if k in allowed_fields}
+            if "Attributes" in cleaned and isinstance(cleaned["Attributes"], dict):
+                cleaned["Attributes"] = remove_all_quotes(cleaned["Attributes"])
+            if "Children" in cleaned and isinstance(cleaned["Children"], list):
+                cleaned["Children"] = [sanitize_rule(c) for c in cleaned["Children"]]
+            if "Actions" in cleaned and isinstance(cleaned["Actions"], list):
+                actions = []
+                for action in cleaned["Actions"]:
+                    a = {"ActionName": action.get("ActionName")}
+                    if "ChildRules" in action and isinstance(action["ChildRules"], list):
+                        a["ChildRules"] = [sanitize_rule(cr) for cr in action["ChildRules"]]
+                    actions.append(a)
+                cleaned["Actions"] = actions
+            return cleaned
+
+        sanitized = [sanitize_rule(r) for r in data]
+        add_missing_guids_if_needed(sanitized)
+        LOGGER.info("JSON file %s successfully loaded", filepath)
+        return sanitized
+    except (json.JSONDecodeError, FileNotFoundError, ValueError) as exc:
+        LOGGER.error("Error loading JSON file %s: %s", filepath, exc)
+        return None
+
+# ---------------------------------------------------------------------------
+# File helpers
+# ---------------------------------------------------------------------------
+
+def allowed_file(filename: str) -> bool:
+    """Return True if ``filename`` has an approved extension."""
+    allowed_exts = {"json", "mmd"}
+    return "." in filename and filename.rsplit(".", 1)[1].lower() in allowed_exts
+
+
+def ensure_directory_exists(directory: str | Path) -> None:
+    """Ensure ``directory`` exists, creating it if necessary."""
+    path = Path(directory)
+    if not path.exists():
+        try:
+            path.mkdir(parents=True, exist_ok=True)
+            LOGGER.info("Created directory: %s", directory)
+        except OSError as exc:
+            LOGGER.error("Failed to create directory %s: %s", directory, exc)
+            raise
+
+
+def get_file_metadata(filepath: str | Path) -> Dict[str, float]:
+    """Return basic file metadata for ``filepath``."""
+    stat = Path(filepath).stat()
+    return {"size": stat.st_size, "last_modified": stat.st_mtime}
+
+# ---------------------------------------------------------------------------
+# Rule helpers
+# ---------------------------------------------------------------------------
+
+def get_dynamic_groups(diagrams_folder: str | Path) -> List[Dict[str, Any]]:
+    """Return groups of diagrams by common prefix."""
+    if not Path(diagrams_folder).exists():
+        return []
+    grouped_catalog: Dict[str, Dict[str, Any]] = {}
+    for folder in sorted(os.listdir(diagrams_folder)):
+        folder_path = Path(diagrams_folder) / folder
+        if not folder_path.is_dir():
+            continue
+        parts = folder.split("_")
+        prefix = "_".join(parts[:2]) if len(parts) > 1 else folder
+        grouped_catalog.setdefault(prefix, {"category": prefix, "entries": []})
+        for filename in os.listdir(folder_path):
+            if filename.endswith(".mmd") or filename.endswith(".json"):
+                existing_entry = next(
+                    (e for e in grouped_catalog[prefix]["entries"] if e["root"] == folder),
+                    None,
+                )
+                if not existing_entry:
+                    grouped_catalog[prefix]["entries"].append(
+                        {
+                            "root": folder,
+                            "diagram": filename if filename.endswith(".mmd") else None,
+                            "hierarchy": filename if filename.endswith(".json") else None,
+                        }
+                    )
+                else:
+                    if filename.endswith(".mmd"):
+                        existing_entry["diagram"] = filename
+                    elif filename.endswith(".json"):
+                        existing_entry["hierarchy"] = filename
+    groups = list(grouped_catalog.values())
+    for g in groups:
+        g["entries"].sort(key=lambda e: e["root"])
+    return sorted(groups, key=lambda g: g["category"])
+
+
+def flatten_rules(rules: Iterable[Dict[str, Any]]) -> List[Dict[str, Any]]:
+    """Flatten a nested structure of rules into a list."""
+    flat_list: List[Dict[str, Any]] = []
+    seen = set()
+
+    def dfs(rule, parent_guid=None, parent_action_index=None):
+        guid = rule.get("RuleGUID")
+        if not guid:
+            guid = f"udf_{uuid.uuid4()}"
+            rule["RuleGUID"] = guid
+        rule["ParentGUID"] = parent_guid
+        rule["ParentActionIndex"] = parent_action_index
+        if guid in seen:
+            return
+        seen.add(guid)
+        flat_list.append(rule)
+        for child in rule.get("Children", []):
+            dfs(child, guid, None)
+        for i, action in enumerate(rule.get("Actions", [])):
+            for child_rule in action.get("ChildRules", []):
+                dfs(child_rule, guid, i)
+
+    for r in rules:
+        dfs(r, parent_guid=None)
+    return flat_list
+
+
+def propagate_disabled_rules(rules: Iterable[Dict[str, Any]], inherited_disabled: bool = False) -> None:
+    """Propagate the disabled state through the rules in-place."""
+    for rule in rules:
+        current_disabled = inherited_disabled or (
+            rule.get("Attributes", {}).get("_Disabled") in ["1", 1, True]
+        )
+        if "Attributes" not in rule:
+            rule["Attributes"] = {}
+        if current_disabled:
+            rule["Attributes"]["_Disabled"] = "1"
+        if "Children" in rule:
+            propagate_disabled_rules(rule["Children"], current_disabled)
+        if "Actions" in rule:
+            for action in rule.get("Actions", []):
+                if "ChildRules" in action:
+                    propagate_disabled_rules(action["ChildRules"], current_disabled)
+
+
+def build_edge_map(edges: Iterable[Dict[str, Any]]) -> Dict[str, list[str]]:
+    """Build a mapping of edges for quick access."""
+    edge_map: Dict[str, list[str]] = {}
+    for edge in edges:
+        parts = edge["edge_str"].split("-->")
+        if len(parts) == 2:
+            left = parts[0].strip()
+            right = parts[1].split("|")[-1].strip()
+            edge_map.setdefault(left, []).append(right)
+    return edge_map
+
+
+def build_all_edges(rules):
+    """Return empty edges list placeholders for tests."""
+    return [], []
+
+
+def build_nodes(*args, **kwargs):
+    """Return an empty node mapping placeholder for tests."""
+    return {}
+
+
+def validate_hierarchy_data(data):
+    """Validate and normalize hierarchy data structure."""
+    if isinstance(data, dict):
+        rules = data.get("rules", [data])
+    else:
+        rules = data
+    normalized = []
+    guid_map = {}
+    for rule in rules:
+        if not isinstance(rule, dict):
+            continue
+        rule["RuleGUID"] = rule.get("RuleGUID") or rule.get("_RuleGUID") or str(uuid.uuid4())
+        rule["RuleName"] = rule.get("RuleName") or rule.get("_RuleName") or f"Rule_{rule['RuleGUID'][:8]}"
+        rule.pop("_RuleGUID", None)
+        rule.pop("_RuleName", None)
+        guid_map[rule["RuleGUID"]] = rule
+        normalized.append(rule)
+    return normalized, guid_map
+
+
+def build_hierarchy(normalized_rules, guid_map):
+    """Build proper hierarchy structure from normalized rules."""
+    roots = []
+    for rule in normalized_rules:
+        parent_guid = rule.get("ParentGUID")
+        if parent_guid and parent_guid in guid_map:
+            parent = guid_map[parent_guid]
+            parent.setdefault("children", []).append(rule)
+        else:
+            roots.append(rule)
+    return roots
+
+# ---------------------------------------------------------------------------
+# Diagram helpers
+# ---------------------------------------------------------------------------
+
+@dataclass
+class DiagramInfo:
+    filename: str
+    created: str
+
+
+def diagram_type_from_filename(filename: str) -> str | None:
+    """Return the diagram type inferred from ``filename``."""
+    fname = filename.lower()
+    if "flowchart" in fname:
+        return "flowchart"
+    if "sequence" in fname:
+        return "sequence"
+    if "class" in fname:
+        return "class"
+    return None
+
+
+def generate_files(json_data: list, output_dir: str) -> list[DiagramInfo]:
+    """Process ``json_data`` and generate diagram and JSON files."""
+    LOGGER.info("Processing %d rules...", len(json_data))
+    os.makedirs(output_dir, exist_ok=True)
+    propagate_disabled_rules(json_data)
+    flat_rules = flatten_rules(json_data)
+    edges, _ = build_all_edges(json_data)
+    edge_map = build_edge_map(edges)
+    LOGGER.info(
+        "Processed %d flat rules and %d edges.",
+        len(flat_rules),
+        len(edges),
+    )
+    container_groups: Dict[str, list] = {}
+    for rule in flat_rules:
+        container = rule.get("Container") or os.path.basename(output_dir)
+        container_groups.setdefault(container, []).append(rule)
+    root_name = os.path.basename(output_dir)
+    infos: list[DiagramInfo] = []
+    for container, rules in container_groups.items():
+        group_ids = {r["RuleGUID"] for r in rules}
+        group_nodes = build_nodes(
+            rules,
+            group_ids=group_ids,
+            root_name=root_name,
+            edge_map=edge_map,
+            rule_to_page=None,
+            current_page=1,
+        )
+        group_edges = [
+            e for e in edges if e["edge_str"].split("-->")[0].strip() in group_ids
+        ]
+        mermaid_code = generate_mermaid_code(group_nodes, group_edges, layout="TD")
+        sanitized_container = secure_filename(container)
+        diagram_filename = f"{sanitized_container}.mmd"
+        diagram_path = os.path.join(output_dir, diagram_filename)
+        with open(diagram_path, "w", encoding="utf-8") as f:
+            f.write(mermaid_code)
+        info = DiagramInfo(diagram_filename, datetime.now(timezone.utc).isoformat())
+        infos.append(info)
+        LOGGER.info("Created diagram %s", info.filename)
+        with open(os.path.join(output_dir, f"{sanitized_container}.json"), "w", encoding="utf-8") as f:
+            json.dump({"rules": rules}, f, indent=4)
+    return infos
+
+
+def generate_mermaid_code(nodes: Dict[str, Dict[str, str]], edges: Iterable[Dict[str, str]], layout: str = "TD") -> str:
+    """Convert nodes and edges into a simple mermaid diagram string."""
+    lines = [f"graph {layout}"]
+    for node_id, node in nodes.items():
+        lines.append(f"{node_id}[{node['label']}]")
+    for edge in edges:
+        lines.append(edge["edge_str"])
+    return "\n".join(lines)
+
+# ---------------------------------------------------------------------------
+# Logging helpers
+# ---------------------------------------------------------------------------
+
+@dataclass
+class ActivityLogEntry:
+    action: str
+    user: str
+    details: str
+    timestamp: str = dataclasses.field(default_factory=lambda: datetime.now(timezone.utc).isoformat())
+
+
+def log_activity(action: str, rule_id: str | None = None, user: str | None = None, details: str | None = None) -> None:
+    """Log an activity entry to ``Config.ACTIVITY_LOG``."""
+    Config.ensure_data_dir()
+    entry = ActivityLogEntry(
+        action=action,
+        user=user or "anonymous",
+        details=details or f"{action} operation",
+    )
+    try:
+        data: dict = {}
+        if os.path.exists(Config.ACTIVITY_LOG):
+            with open(Config.ACTIVITY_LOG, "r", encoding="utf-8") as f:
+                data = json.load(f)
+        else:
+            data = {"rules": {}, "activity_log": []}
+        data.setdefault("activity_log", []).append(asdict(entry))
+        if rule_id:
+            data.setdefault("rules", {})[rule_id] = {
+                "status": "active",
+                "last_modified": entry.timestamp,
+                "modified_by": entry.user,
+            }
+        with open(Config.ACTIVITY_LOG, "w", encoding="utf-8") as f:
+            json.dump(data, f, indent=2)
+    except Exception as exc:  # pragma: no cover - log failures shouldn't crash
+        LOGGER.error("Failed to log activity: %s", exc)
+
+
+def get_rule_stats() -> dict:
+    """Return rule counts and recent activity totals."""
+    Config.ensure_data_dir()
+    try:
+        with open(Config.ACTIVITY_LOG, "r", encoding="utf-8") as f:
+            data = json.load(f)
+        now = datetime.utcnow()
+        totals = {
+            "total_rules": len(data.get("rules", {})),
+            "last_7_days": 0,
+            "last_30_days": 0,
+            "last_90_days": 0,
+        }
+        for entry in data.get("activity_log", []):
+            try:
+                ts = datetime.fromisoformat(entry["timestamp"])
+            except Exception:
+                continue
+            if ts >= now - timedelta(days=7):
+                totals["last_7_days"] += 1
+            if ts >= now - timedelta(days=30):
+                totals["last_30_days"] += 1
+            if ts >= now - timedelta(days=90):
+                totals["last_90_days"] += 1
+        return totals
+    except Exception as exc:  # pragma: no cover - unexpected errors
+        LOGGER.error("Failed to compute stats: %s", exc)
+        return {"total_rules": 0, "last_7_days": 0, "last_30_days": 0, "last_90_days": 0}
+
+
+def get_activity_trend(days: int = 30) -> list[dict]:
+    """Return daily activity counts for the past ``days`` days."""
+    Config.ensure_data_dir()
+    try:
+        with open(Config.ACTIVITY_LOG, "r", encoding="utf-8") as f:
+            data = json.load(f)
+        now = datetime.utcnow()
+        start_date = now.date() - timedelta(days=days - 1)
+        counts = { (start_date + timedelta(days=i)).isoformat(): 0 for i in range(days) }
+        for entry in data.get("activity_log", []):
+            try:
+                ts = datetime.fromisoformat(entry["timestamp"]).date()
+            except Exception:
+                continue
+            if ts.isoformat() in counts:
+                counts[ts.isoformat()] += 1
+        return [
+            {"label": d, "count": counts[d]} for d in sorted(counts.keys())
+        ]
+    except Exception as exc:  # pragma: no cover - unexpected errors
+        LOGGER.error("Trend calculation failed: %s", exc)
+        return []
+
+# ---------------------------------------------------------------------------
+# Misc helpers
+# ---------------------------------------------------------------------------
+
+def highlight_matches(text: str, query: str) -> str:
+    """Highlight matches case-insensitively."""
+    if not query:
+        return text
+    try:
+        words: List[str] = [w for w in query.split() if w]
+        if len(words) > 1:
+            for w in words:
+                pattern = re.compile(re.escape(w), re.IGNORECASE)
+                text = pattern.sub(lambda m: f"<strong>{m.group()}</strong>", text)
+            return text
+        pattern = re.compile(re.escape(query), re.IGNORECASE)
+        return pattern.sub(lambda m: f"<strong>{m.group()}</strong>", text)
+    except re.error as exc:  # pragma: no cover - regex errors
+        LOGGER.error("Invalid regex in highlight_matches: %s", exc)
+        return text
+
+
+def get_snippet(content: str, query: str, snippet_length: int = 100) -> str:
+    """Extract a short snippet of ``content`` around ``query`` without cutting words."""
+    index = content.lower().find(query.lower())
+    if index == -1:
+        return ""
+    if len(content) <= len(query) + snippet_length:
+        return content
+    half = snippet_length // 2
+    start = max(0, index - half)
+    start = content.rfind(" ", 0, start)
+    start = 0 if start == -1 else start + 1
+    end = min(len(content), index + len(query) + half)
+    space_after = content.rfind(" ", index + len(query), end)
+    if space_after != -1 and space_after > index + len(query):
+        end = space_after
+    snippet = content[start:end]
+    return snippet.strip()
+
+
+def get_current_user() -> str:
+    """Return the username of the authenticated user or 'anonymous'."""
+    try:
+        return current_user.username if current_user.is_authenticated else "anonymous"
+    except Exception:  # pragma: no cover - runtime guard
+        return "anonymous"
+
+
+def initialize_directories(app) -> None:
+    """Ensure required application directories exist."""
+    required_dirs = {
+        "uploads": app.config.get("UPLOAD_FOLDER", "./uploads"),
+        "diagrams": app.config.get("DIAGRAMS_FOLDER", "./diagrams"),
+        "help": os.path.join(app.static_folder, "help"),
+    }
+    for path in required_dirs.values():
+        try:
+            os.makedirs(path, exist_ok=True)
+            app.logger.info("Directory verified: %s", path)
+        except OSError as e:  # pragma: no cover - unexpected
+            app.logger.error("Failed to create directory %s: %s", path, e)
+            raise
+
+
+def get_help_topics() -> list[str]:
+    """Return a sorted list of available help topic names."""
+    help_dir = Path(current_app.root_path) / "static" / "help"
+    if not help_dir.exists():
+        return []
+    return sorted(f.stem for f in help_dir.glob("*.md"))
+
+
+__all__ = [
+    "allowed_file",
+    "ensure_directory_exists",
+    "load_and_sanitize_json",
+    "generate_files",
+    "log_activity",
+    "diagram_type_from_filename",
+    "get_snippet",
+    "get_current_user",
+    "initialize_directories",
+    "get_help_topics",
+    "get_dynamic_groups",
+    "flatten_rules",
+    "propagate_disabled_rules",
+    "build_edge_map",
+    "validate_hierarchy_data",
+    "build_hierarchy",
+    "ActivityLogEntry",
+    "remove_all_quotes",
+    "add_missing_guids_if_needed",
+    "get_file_metadata",
+    "highlight_matches",
+    "get_rule_stats",
+    "get_activity_trend",
+]
